+++ apps/docs/src/components/feedbacks/ContextualHelp.stories.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-import { Alert, RoundButton } from "@react-elf/ui";
 
 const ContextualHelp = () => {
   return <div></div>;
 };
 
 const meta = {
+  title: "Components/Feed backs/Contextual Help",
   component: ContextualHelp,
   tags: ["autodocs"],
 } satisfies Meta<typeof ContextualHelp>;