+++ packages/@react-elf-types/tools/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @react-elf-types/tools
 
+## 0.0.81
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf-types/tooltip@0.0.81
+
 ## 0.0.76
 
 ### Patch Changes