+++ packages/@react-elf/button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @react-elf/button
 
+## 0.0.81
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf/tooltip@0.0.81
+
 ## 0.0.76
 
 ### Patch Changes