+++ packages/@react-elf/toolbar/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @react-elf/toolbar
 
+## 0.0.81
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf/tools@0.0.81
+  - @react-elf-types/toolbar@0.0.81
+
 ## 0.0.76
 
 ### Patch Changes