+++ packages/@react-elf/tooltip/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @react-elf/button
 
+## 0.0.81
+
+### Patch Changes
+
+- add Tooltip mdx
+
+- Updated dependencies []:
+  - @react-elf-types/tooltip@0.0.81
+
 ## 0.0.76
 
 ### Patch Changes